# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank Card Management API",
    "description": "API for managing bank cards and transfers",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/transfers": {
      "post": {
        "tags": [
          "Transfers"
        ],
        "summary": "Transfer money between user's own cards",
        "description": "Only USER can transfer money and only between own cards",
        "operationId": "transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Transfer completed successfully"
          },
          "400": {
            "description": "Invalid input data or validation failed"
          },
          "401": {
            "description": "Unauthorized - JWT token missing, expired or invalid"
          },
          "403": {
            "description": "Forbidden - user cannot transfer between cards they do not own"
          },
          "404": {
            "description": "Card not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferRequest": {
        "type": "object",
        "properties": {
          "fromCardId": {
            "type": "integer",
            "format": "int64"
          },
          "toCardId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "amount"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}